Quote all arguments, not just import paths
When the path .capnp file passed to the compiler contains spaces (e.g.
`C:\Users\Full Name\source\repos\...`) the argument was not quoted
causing it to be misinterpreted as multiple arguments.

.net sdk 5.0 adds a new ArgumentsList property to ProcessStartInfo which
can handle more edge cases. This has been tested but not implemented in
order to avoid changing the projects, instead this commit only moves the
rudimentary quoting already applied to import paths so that it applies
to all arguments.